Address the -Wsign-compare compiler warnings.

+++ lib/proxy/ftp/data.c
@@ -1,6 +1,6 @@
 /*
  * ProFTPD - mod_proxy FTP data conn routines
- * Copyright (c) 2012-2015 TJ Saunders
+ * Copyright (c) 2012-2016 TJ Saunders
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -132,7 +132,7 @@ int proxy_ftp_data_send(pool *p, conn_t *data_conn, pr_buffer_t *pbuf,
   pr_timer_reset(PR_TIMER_STALLED, ANY_MODULE);
   pr_timer_reset(PR_TIMER_IDLE, ANY_MODULE);
 
-  if (nwrote == pbuf->remaining) {
+  if ((size_t) nwrote == pbuf->remaining) {
     pbuf->current = NULL;
     pbuf->remaining = 0;
